requirements.txt
numpy
keras
absl-py
pillow
sklearn
opencv-python
